MNT: Move functions from future to their proper location (Fixes #1112)
This makes their behavior standard and eliminates the old behavior.
